requirements.txt
Django==2.1
django-annoying==0.10.4
django-autocomplete-light==3.3.0
Pillow==5.2.0
psycopg2-binary==2.7.5
python-barcode==0.9.0
pytz==2018.5
